+++ styles/styleTree/chatPanel.ts
